requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
watchdog>=3.0.0
